Reused depends in checkdeps script and formatted it also
